Package refactoring and Motif reporter class
