+++ src/main/java/ru/spbstu/calculator/Constants.java
@@ -2,4 +2,6 @@ package ru.spbstu.calculator;
 
 public class Constants {
     public static final int DEFAULT_WINDOW_SIZE = 10;
+    public static final Character[] NUCLEOTIDES = {'A', 'C', 'G', 'T'};
+    public static final double frequency = 0.25f;
 }
