--- src/main/java/ru/spbstu/pwmatrix/Constants.java
@@ -1,5 +1,0 @@
-package ru.spbstu.pwmatrix;
-
-public class Constants {
-    public static final Character[] NUCLEOTIDES = {'A', 'C', 'G', 'T'};
-}
