+++ src/main/resources/PWMSample
@@ -1,0 +1,30 @@
+>ce1cg 17 61
+GAGGTAAACTAATGTTTGTGCTGGTTTTTGTGGCATCGGGCGAGAATAGCGCGTGGTGTGAAAGACTGTTTTTTTGATCGTTTTCACAA
+AAATGGAAGTCCACAGTCTTGACAG
+>ara 17 55
+TCCGTAAGTGACAAAAACGCGTAACAAAAGTGTCTATAATCACGGCAGAAAAGTCCACATTGATTATTTGCACGGCGTCACACTTTGCT
+ATGCCATAGCATTTTTATCCATAAG
+>bglr1 76
+CAGGTTGGAACAAATCCCAATAACTTAATTATTGGGATTTGTTATATATAACTTTATAAATTCCTAAAATTACACAAAGTTAATAACTG
+TGAGCATGGTCATATTTTTATCAAT
+>crp 63
+ACAGTCAGTCACAAAGCGAAAGCTATGCTAAAACAGTCAGGATGCTACAGTAATACATTGATGTACTGCATGTATGCAAAGGACGTCAC
+ATTACCGTGCAGTACAGTTGATAGC
+>cya 50
+TAGGTCATTACGGTGCTACACTTGTATGTAGCGCATCTTTCTTTACGGTCAATCAGCAAGGTGTTAAATTGATCACGTTTTAGACCATT
+TTTTCGTCGTGAAACTAAAAAAACC
+>deop2 7 60
+TAGGTACTGAGTGAATTATTTGAACCAGATCGCATTACAGTGATGCAAACTTGTAAGTAGATTTCCTTAATTGTGATGTGTATCGAAGT
+GTGTTGCGGAGTAGATGTTAGAATA
+>gale 42
+ATGGTAACTGCGCATAAAAAACGGCTAAATTCTTGTGTAAACGATTCCACTAATTTATTCCATGTCACACTTTTCGCATCTTTGTTATG
+CTATGGTTATTTCATACCATAAGCC
+>ilv 39
+CAGGTATACGCTCCGGCGGGGTTTTTTGTTATCTGCAATTCAGTACAAAACGTGATCAACCCCTCAATTTTCCCTTTGCTGAAAAATTT
+TCCATTGTCTCCCCTGTAAAGCTGT
+>lac 9 80
+TGTGTGAGTAACGCAATTAATGTGAGTTAGCTCACTCATTAGGCACCCCAGGCTTTACACTTTATGCTTCCGGCTCGTATGTTGTGTGG
+AATTGTGAGCGGATAACAATTTCAC
+>male 14
+AAGGTAAGTACATTACCGCCAATTCTGTAACAGAGATCACACAAAGCGACGGTGGGGCGTAGGGGCAAGGAGGATGGAAAGAGGTTGCC
+GTATAAAGAAACTAGAGTCCGTTTA
